Removes obsolete extension and moves settings to serializer
